+++ commands/auto/role.js
@@ -6,7 +6,7 @@ exports.run = async (client, interaction) => {
         const sendMessage = require('../../util/sendMessage');
         const Discord = require("discord.js");
         const isAdmin = require("../../util/isAdmin");
-        const { EligibleRoles } = require('../../database/dbServices/attatchments.api');
+        const { EligibleRoles } = require('../../database/dbServices/server.api');
 
         let ephemeral = true;
         let ephemeralArg = interaction.options.getBoolean("ephemeral");
