+++ events/guildMemberUpdate.js
@@ -4,7 +4,7 @@ module.exports = async (client, member, newMember) => {
     let globalVars = require('./ready');
     try {
         const Discord = require("discord.js");
-        const { LogChannels } = require('../database/dbServices/attatchments.api');
+        const { LogChannels } = require('../database/dbServices/server.api');
 
         let logChannel = await LogChannels.findOne({ where: { server_id: member.guild.id } });
         if (!logChannel) return;
@@ -65,7 +65,7 @@ module.exports = async (client, member, newMember) => {
                 if (e.toString().includes("Missing Permissions")) executor = null;
             };
 
-            const { PersonalRoles, PersonalRoleServers } = require('../database/dbServices/attatchments.api');
+            const { PersonalRoles, PersonalRoleServers } = require('../database/dbServices/server.api');
             let serverID = await PersonalRoleServers.findOne({ where: { server_id: member.guild.id } });
             let roleDB = await PersonalRoles.findOne({ where: { server_id: member.guild.id, user_id: member.id } });
             if (!newMember.premiumSince && serverID && roleDB && !member.permissions.has("MANAGE_ROLES")) await deleteBoosterRole();
