+++ events/ready.js
@@ -4,7 +4,7 @@ module.exports = async (client) => {
         const stan = require('../affairs/stan')(client);
         const birthday = require('../affairs/birthday')(client);
         // const { bank } = require('../database/bank');
-        // const { Users } = require('../database/dbServices/attatchments.api');
+        // const { Users } = require('../database/dbServices/server.api');
         // const storedBalances = await Users.findAll();
         // storedBalances.forEach(b => bank.currency.set(b.user_id, b));
         const getTime = require('../util/getTime');
